applied review feedback

+++ scripts/Set-AzureAdApplication/v2/Set-AadApplication.psm1
@@ -151,36 +151,38 @@ function Set-AadApplication {
             }
         }
 
-        # Check for existing secrets and remove them so they can be re-created
-        Write-Verbose "Checking for existing secrets"
-        $appKeySecrets = Get-AzureADApplicationPasswordCredential -ObjectId $application.ObjectId
+        if($Secrets){ 
+            # Check for existing secrets and remove them so they can be re-created
+            Write-Verbose "Checking for existing secrets"
+            $appKeySecrets = Get-AzureADApplicationPasswordCredential -ObjectId $application.ObjectId
 
-        if($appKeySecrets)  {
-            foreach($existingSecret in $appKeySecrets) {
-                foreach($secret in $Secrets) {
-                    if([System.Text.Encoding]::ASCII.GetString($existingSecret.CustomKeyIdentifier) -eq $secret.Description) {
+            if($appKeySecrets)  {
+                foreach($existingSecret in $appKeySecrets) {
+                    foreach($secret in $Secrets) {
+                        if([System.Text.Encoding]::ASCII.GetString($existingSecret.CustomKeyIdentifier) -eq $secret.Description) {
 
-                        $stringDescription = $secret.Description | Out-String
-                        $trimmedStringDescription = $stringDescription -replace [Environment]::NewLine,"";
+                            $stringDescription = $secret.Description | Out-String
+                            $trimmedStringDescription = $stringDescription -replace [Environment]::NewLine,"";
 
-                        Write-Verbose "Removing existing key with description: $trimmedStringDescription"
-                        Remove-AzureADApplicationPasswordCredential  -ObjectId $application.ObjectId -KeyId $existingSecret.KeyId
+                            Write-Verbose "Removing existing key with description: $trimmedStringDescription"
+                            Remove-AzureADApplicationPasswordCredential  -ObjectId $application.ObjectId -KeyId $existingSecret.KeyId
+                        }
                     }
                 }
             }
-        }
 
-        # Create new secrets
-        foreach($secret in $Secrets) {
-            $endDate = [datetime]::ParseExact($secret.EndDate,'dd/MM/yyyy',[Globalization.CultureInfo]::InvariantCulture)
-            
-            $stringDescription = $secret.Description | Out-String
-            $trimmedStringDescription = $stringDescription -replace [Environment]::NewLine,"";
-            
-            Write-Verbose "Creating new key with description: $trimmedStringDescription and end date $endDate"
-            $appKeySecret = New-AzureADApplicationPasswordCredential -ObjectId $application.ObjectId -CustomKeyIdentifier $secret.Description -EndDate $endDate
-            
-            Write-Host "##vso[task.setvariable variable=Secret.$trimmedStringDescription;isOutput=true]$($appKeySecret.Value)"
+            # Create new secrets
+            foreach($secret in $Secrets) {
+                $endDate = [datetime]::ParseExact($secret.EndDate,'dd/MM/yyyy',[Globalization.CultureInfo]::InvariantCulture)
+                
+                $stringDescription = $secret.Description | Out-String
+                $trimmedStringDescription = $stringDescription -replace [Environment]::NewLine,"";
+                
+                Write-Verbose "Creating new key with description: $trimmedStringDescription and end date $endDate"
+                $appKeySecret = New-AzureADApplicationPasswordCredential -ObjectId $application.ObjectId -CustomKeyIdentifier $secret.Description -EndDate $endDate
+                
+                Write-Host "##vso[task.setvariable variable=Secret.$trimmedStringDescription;isOutput=true]$($appKeySecret.Value)"
+            }
         }
 
         Write-Information "Owners of the application are now:"
